+++ ichnaea/service/monitor/views.py
@@ -1,5 +1,6 @@
 import time
 
+from pyramid.httpexceptions import HTTPServiceUnavailable
 from pyramid.view import view_config
 
 
@@ -35,6 +36,13 @@ def check_database(request):
     return _check_timed(request.db_slave_session.ping)
 
 
+def check_geoip(request):
+    geoip_db = request.registry.geoip_db
+    result = _check_timed(geoip_db.ping)
+    result['age_in_days'] = geoip_db.age
+    return result
+
+
 def check_redis(request):
     return _check_timed(request.registry.redis_client.ping)
 
@@ -47,16 +55,27 @@ def check_stats(request):
 def monitor_view(request):
     services = {
         'database': check_database,
+        'geoip': check_geoip,
         'redis': check_redis,
         'stats': check_stats,
     }
+    failed = False
     result = {}
     for name, check in services.items():
         try:
             service_result = check(request)
         except Exception:  # pragma: no cover
-            result[name] = {'up': False}
+            result[name] = {'up': None, 'time': -1}
+            failed = True
         else:
             result[name] = service_result
+            if not service_result['up']:
+                failed = True
+
+    if failed:
+        response = HTTPServiceUnavailable()
+        response.content_type = 'application/json'
+        response.json = result
+        return response
 
     return result