+++ CHANGES.rst
@@ -14,6 +14,7 @@ Migrations
 Changes
 ~~~~~~~
 
+- #274: Extend monitor view to include geoip db status.
 
 20150127130000
 **************